[feature] split key method: basic && sugar
